[Refactor] class명 변경, collection view layout 변경
